fix: refactor MS Build 2026 workflow deployment and agent creation
- deploy_msbuild_workflow.py: strip id: line from YAML, delete existing workflows by ID, POST with new {id, yaml} structure to fix poisoned registry state
- create_msbuild_agent.py: fix param types (long → integer) for threshold_us ES|QL tool parameter
- elastic-agent-pr-review.yaml: add id field for workflow identification
- Update backend services, frontend components, and .env.example for MS Build demo integration
- Add draft documentation and session bucket files for workshop planning

This completes Step 1-2 of MS Build 2026 Block 1 deployment prep.

+++ scripts/create_msbuild_agent.py
@@ -164,7 +164,7 @@ def main() -> int:
         ),
         params={
             "endpoint": {"type": "string", "description": "HTTP route to match"},
-            "threshold_us": {"type": "long", "description": "Microsecond threshold; 3000000 = 3 seconds"},
+            "threshold_us": {"type": "integer", "description": "Microsecond threshold; 3000000 = 3 seconds"},
         },
     )
     if tid:
